JioTV Added As Seperate Route
