+++ templates/JioTV/player.html
@@ -8,7 +8,7 @@
       href="https://cdnjs.cloudflare.com/ajax/libs/normalize/5.0.0/normalize.min.css"
     />
     <link rel="stylesheet" href="https://unpkg.com/plyr@3/dist/plyr.css" />
-    <link rel="stylesheet" href="static/css/style.css" />
+    <link rel="stylesheet" href="/static/JioTV/css/style.css" />
   </head>
   <body>
     <!-- partial:index.partial.html -->
@@ -27,6 +27,6 @@
     <!-- partial -->
     <script src="https://cdn.polyfill.io/v2/polyfill.min.js?features=es6,Array.prototype.includes,CustomEvent,Object.entries,Object.values,URL"></script>
     <script src="https://unpkg.com/plyr@3"></script>
-    <script src="static/js/script.js"></script>
+    <script src="/static/JioTV/js/script.js"></script>
   </body>
 </html>
